fix(metrics): satisfy strict product:lint by moving api.py into presentation/
The metrics product is in strict isolation mode (it has backend/facade/
contracts.py), so hogli product:lint flags backend/api.py at root as
misplaced — the canonical location is backend/presentation/api.py.

- Move backend/api.py -> backend/presentation/api.py
- Update the import in posthog/api/__init__.py
- Add the matching tach interface block exposing
  backend.facade and backend.presentation.api
- Register products.metrics under posthog's depends_on so the import
  is allowed by tach
- Update the report_user_action mock path in the test

Generated-By: PostHog Code
Task-Id: c20a4a56-d258-420f-8b00-df6aaf3083b5
